+++ number/unit.go
@@ -7,9 +7,6 @@ import (
 
 type Unit int64
 
-// Deprecated:  the type `One` will deprecated when `v1.1.0` please use `Unit` replace.
-type One = Unit
-
 // 1k = 1000 thousand/kilo
 // 1M = 1000_000 million
 // 1G = 1000_000_000 billion